+++ internal/testreadme/main_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/interline-io/transitland-lib/copier"
 	"github.com/interline-io/transitland-lib/tlcsv"
 	"github.com/interline-io/transitland-lib/tldb"
-	_ "github.com/interline-io/transitland-lib/tldb/tlsqlite"
+	_ "github.com/interline-io/transitland-lib/tldb/sqlite"
 )
 
 // var URL = "https://github.com/interline-io/transitland-lib/raw/master/testdata/external/bart.zip"
